mohsen_backend_gateway/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
python-dotenv==0.19.2
Werkzeug==2.0.2
kafka-python==2.0.2
msgpack==1.0.3
flask_swagger_ui==3.36.0
requests==2.26.0
pymongo==4.0.1
jsonschema==4.3.0
redis==4.0.2
flask-cors==3.0.10
PyJWT==2.3.0

